Generate correct "type" for Integer fields
Using a "type" of "number" and a "format" of "integer" is not documented
in the JSON schema specification as a correct way to represent integers.
Instead, a "type" of "integer" should be used as this ensure that
validators will correctly validate inputs as an integer and not accept
"number" values.

https://json-schema.org/understanding-json-schema/reference/numeric.html?highlight=integer#integer

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>